size t type for external sort

+++ src/enclave/App/App.cpp
@@ -1196,7 +1196,7 @@ Java_edu_berkeley_cs_rise_opaque_execution_SGXEnclave_EnclaveColumnSort(
     env->GetByteArrayElements(input_rows, &if_copy));
 
   uint8_t *output_buffer = (uint8_t *) malloc(sizeof(uint8_t) * r);
-  uint32_t output_buffer_size;
+  size_t output_buffer_size;
 
   if (round == 0) {
     sgx_check("Column Sort Pad", ecall_column_sort_pad(eid, input_rows_ptr, input_length, r, s, &output_buffer, &output_buffer_size));
